requirements.txt
py-solc-x==1.1.1 
Flask==2.1.2